+++ IsekaiMod/Content/Backgrounds/MartialArtist.cs
@@ -1,4 +1,5 @@
-﻿using Kingmaker.Blueprints;
+﻿using IsekaiMod.Utilities;
+using Kingmaker.Blueprints;
 using Kingmaker.Blueprints.Classes;
 using Kingmaker.Enums;
 using Kingmaker.UnitLogic.FactLogic;
@@ -14,10 +15,7 @@ namespace IsekaiMod.Content.Backgrounds {
             var ExoticWeaponProficiency = BlueprintTools.GetModBlueprint<BlueprintFeature>(IsekaiContext, "ExoticWeaponProficiency");
             var BackgroundMartialArtist = Helpers.CreateBlueprint<BlueprintFeature>(IsekaiContext, "BackgroundMartialArtist", bp => {
                 bp.SetName(IsekaiContext, "Martial Artist");
-                bp.SetDescription(IsekaiContext, "The Martial Artist is proficient with all exotic weapons.\n"
-                    + "If the character already has the class skill, {g|Encyclopedia:Weapon_Proficiency}weapon proficiency{/g} or armor proficiency granted by the selected background "
-                    + "from her class during character creation, then the corresponding {g|Encyclopedia:Bonus}bonuses{/g} from background change to a +1 competence bonus in case of skills, "
-                    + "a +1 enhancement bonus in case of weapon proficiency and a -1 Armor {g|Encyclopedia:Check}Check{/g} {g|Encyclopedia:Penalty}Penalty{/g} reduction in case of armor proficiency.");
+                bp.SetBackgroundDescription(IsekaiContext, "The Martial Artist is proficient with all exotic weapons.");
                 bp.AddComponent<AddFacts>(c => {
                     c.m_Facts = new BlueprintUnitFactReference[] { ExoticWeaponProficiency.ToReference<BlueprintUnitFactReference>() };
                 });